+++ docs/source/changelog.rst
@@ -35,6 +35,7 @@ v26.1.12-dev
 - Added :class:`~discord.flags.MessageFlags` flag.
 - Added :class:`~discord.objects.components.CheckboxComponent` component.
 - Added :class:`~discord.objects.components.CheckboxGroupComponent` component.
+- Added :class:`~discord.objects.components.CheckboxGroupOption` object.
 
 **Documentation Changes**